Migrate wdio-allure-reporter to Typescript
