+++ .eslintrc.js
@@ -40,7 +40,7 @@ module.exports = {
         rules: {
             // see https://stackoverflow.com/questions/55280555/typescript-eslint-eslint-plugin-error-route-is-defined-but-never-used-no-un
             'no-unused-vars': 'off',
-            '@typescript-eslint/no-unused-vars': 'error',
+            '@typescript-eslint/no-unused-vars': ['error', { 'argsIgnorePattern': '^_' }],
             'no-undef': 'off'
         }
     }]
