+++ package.json
@@ -63,6 +63,7 @@
   "devDependencies": {
     "@octokit/rest": "^18.0.0",
     "@tsconfig/node10": "^1.0.7",
+    "@types/cheerio": "^0.22.22",
     "@types/jest": "^26.0.0",
     "@types/node": "^12.19.3",
     "@typescript-eslint/eslint-plugin": "^4.4.0",
