+++ packages/devtools/src/index.ts
@@ -1,5 +1,5 @@
 import os from 'os'
-import { UAParser } from 'ua-parser-js'
+import UAParser from 'ua-parser-js'
 import { v4 as uuidv4 } from 'uuid'
 
 import logger from '@wdio/logger'
