+++ packages/wdio-allure-reporter/package.json
@@ -9,7 +9,7 @@
   "engines": {
     "node": ">=10.0.0"
   },
-  "types": "./allure-reporter.d.ts",
+  "types": "./build/index.d.ts",
   "repository": {
     "type": "git",
     "url": "git://github.com/webdriverio/webdriverio.git"
@@ -26,7 +26,9 @@
   "dependencies": {
     "@wdio/reporter": "6.8.1",
     "allure-js-commons": "^1.3.2",
-    "strip-ansi": "^6.0.0"
+    "strip-ansi": "^6.0.0",
+    "@types/allure-js-commons": "^0.0.1",
+    "@types/cucumber": "^6.0.1"
   },
   "peerDependencies": {
     "@wdio/cli": "^6.0.1"
