--- packages/wdio-allure-reporter/tests/__fixtures__/attachment.js
@@ -1,7 +1,0 @@
-export function xmlAttachment() {
-    return {
-        name: 'Page source',
-        type: 'application/xml',
-        content: '<?xml version="1.0" encoding="UTF-8" standalone="yes" ?>'
-    }
-}
