+++ packages/wdio-allure-reporter/tests/__fixtures__/attachment.ts
@@ -1,0 +1,7 @@
+export function xmlAttachment() {
+    return {
+        name: 'Page source',
+        type: 'application/xml',
+        content: '<?xml version="1.0" encoding="UTF-8" standalone="yes" ?>'
+    }
+}
