+++ packages/wdio-allure-reporter/tsconfig.json
@@ -3,7 +3,8 @@
   "compilerOptions": {
     "baseUrl": ".",
     "outDir": "./build",
-    "rootDir": "./src"
+    "rootDir": "./src",
+    "types": ["webdriver", "cucumber"]
   },
   "include": [
     "src/**/*"
