--- scripts/templates/allure-reporter.tpl.d.ts
@@ -1,10 +1,0 @@
-
-declare namespace AllureReporter {
-    type StepStatus = 'passed' | 'failed' | 'broken' | 'canceled' | 'skipped'
-
-    // ... AllureReporter commands ...
-}
-
-declare module "@wdio/allure-reporter" {
-    export default AllureReporter
-}
