+++ scripts/type-generation/generate-all-typings.js
@@ -1,4 +1,3 @@
 require('./webdriver-generate-typings')
 require('./webdriverio-generate-typings')
 require('./devtools-generate-typings')
-require('./allure-reporter-generate-typings')
