+++ tests/typings/webdriverio/tsconfig.json
@@ -3,6 +3,7 @@
     "outDir": "dist",
     "noImplicitAny": true,
     "lib": ["es2018", "dom"],
+    "esModuleInterop": true,
     "types": [
       "webdriverio",
       "@wdio/allure-reporter",
